requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coverage==6.0.2
crcmod==1.7
Django==2.2.4
mistune==0.8.4
django-appconf==1.0.5
django-bootstrap3==14.2.0
django-ckeditor==6.0.0
django-extensions==3.1.3
django-imagekit==4.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-sslserver==0.22
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.3.4
markdown2==2.4.0
openpyxl==3.0.7
pilkit==2.0
Pillow==8.2.0
pycryptodome==3.10.1
Pygments==2.10.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
